refactor(environment): split package manager panels
