+++ src/components/environment/index.ts
@@ -10,6 +10,21 @@ export {
   EnvironmentPackageSummaryPanel,
   type EnvironmentPackageSummaryPanelProps,
 } from "./EnvironmentPackageSummaryPanel";
+export { CondaDependencyPanel, type CondaDependencyPanelProps } from "./CondaDependencyPanel";
+export {
+  DenoDependencyPanel,
+  type DenoConfigInfo,
+  type DenoDependencyPanelProps,
+} from "./DenoDependencyPanel";
+export { PixiDependencyPanel, type PixiDependencyPanelProps } from "./PixiDependencyPanel";
+export {
+  PackageSpecList,
+  parsePackageSpec,
+  type PackageSpecListProps,
+  type PackageSpecTone,
+} from "./PackageSpecList";
+export { UvDependencyPanel, type UvDependencyPanelProps } from "./UvDependencyPanel";
+export { type DependencyPanelVariant, type EnvironmentSyncState } from "./dependency-panel-types";
 export {
   notebookMetadataToPackageViewModel,
   type NotebookPackageManager,
